Comments on complex code finish in property
